fix(boot): Renaming register to compile util
