+++ module/base/src/scan-app.ts
@@ -1,4 +1,5 @@
-import { FsUtil, RegisterUtil, AppCache } from '@travetto/boot';
+import { FsUtil, AppCache } from '@travetto/boot';
+import { FrameworkUtil } from '@travetto/boot/src/framework';
 
 import { Env } from './env';
 import { ScanEntry, ScanFs } from './scan-fs';
@@ -45,7 +46,7 @@ export class ScanApp {
    * Support framework resolution if active
    */
   private static resolveFramework(x: SimpleEntry, root: string) {
-    const file = RegisterUtil.devResolve(x.file);
+    const file = FrameworkUtil.devResolve(x.file);
     return { ...x, file, module: file.replace(`${root}/`, '') };
   }
