+++ module/boot/src-ts/index.ts
@@ -3,5 +3,5 @@ export * from './cache';
 export * from './exec';
 export * from './env';
 export * from './fs';
-export * from './register';
+export * from './compile';
 export * from './transpile';
