+++ module/compiler/src/transpiler.ts
@@ -1,7 +1,9 @@
 import * as ts from 'typescript';
 import * as path from 'path';
 
-import { FileCache, RegisterUtil, TranspileUtil, FsUtil } from '@travetto/boot';
+import { FileCache, TranspileUtil, FsUtil } from '@travetto/boot';
+import { FrameworkUtil } from '@travetto/boot/src/framework';
+
 import { ScanApp } from '@travetto/base';
 import { SystemUtil } from '@travetto/base/src/internal/system';
 
@@ -88,7 +90,7 @@ export class Transpiler {
   private getHost(): ts.CompilerHost {
     const host: ts.CompilerHost = {
       readFile: this.readFile,
-      realpath: x => RegisterUtil.devResolve(x), // @line-if $TRV_DEV
+      realpath: x => FrameworkUtil.devResolve(x), // @line-if $TRV_DEV
       writeFile: this.writeFile,
       fileExists: this.fileExists,
       getDefaultLibFileName: ts.getDefaultLibFileName,
